voucher marking as used issue fixed and sales item model total data type changed

+++ commonapp/models/salesitem.py
@@ -15,7 +15,7 @@ class SalesItem(models.Model):
     rate = models.PositiveIntegerField()
     quantity = models.PositiveIntegerField()
     voucher = models.ForeignKey(Voucher, on_delete=models.PROTECT, null=True, blank=True)
-    total = models.PositiveIntegerField()
+    total = models.DecimalField(max_digits=10, decimal_places=2)
     discount = models.PositiveIntegerField(null=True, blank=True)
     created_at = models.DateTimeField(auto_now_add=True)
